New max that handles edge cases
This now returns nil if an empty list is passed in (rather than
previously returning 0), and also now correctly handles negative
numbers.

Because 0 is not passed in as an initial seed (and instead the head of
the list is), if a list comes through that has all negative numbers, the
largest of those negative numbers will be returned, rather than 0.