Cipher-text to add x characters to single-quoted string
This function takes a single-quoted string (a list of character codes),
and an offset, returning a list of character codes that are each
increased by the offset. It wraps around the alphabet, assumes 1) only
lowercase letters 2) assumes letters only.

I had to look up the ASCII value for z (122) to accomplish this. The
lack of uppercase support right now is a bit of a letdown, but it
satisfies the use case in the exercise...and I really want to keep
reading!