Bail out of ranges where from is higher than to.
This changes the guard clause in my base case of span so that when from
exceeds to, it returns an empty list. The prior implementation assumed
valid input where from never exceeded to. This takes that into account,
and if it encounters it, just short-circuits with an empty list.

This has the nice effect of also serving as a base case for valid input,
because eventually from will exceed to when we're consistently adding 1
to its value, and in that case, we want to provide the empty list so
that we've completed our recursive function, and the rest of the values
unwind to be returned back to the caller.